before ICCE

+++ nhwave_amp/xarray_obj/_input_nc_creation.py
@@ -44,7 +44,7 @@ def ensure_net_cdf_type(nc_data):
     for attr_name, attr_value in nc_data.attrs.items():
         # Convert floats
         if isinstance(attr_value, (float, np.floating)):
-            new_attrs[attr_name] = np.float32(attr_value)
+            new_attrs[attr_name] = np.float64(attr_value)
         # Convert integers
         elif isinstance(attr_value, (int, np.integer)):
             new_attrs[attr_name] = np.int32(attr_value)
@@ -80,6 +80,7 @@ def get_net_cdf(var_dict):
         # Get list of any xarrays (ie- DomainObject, WaveMaker Object)
         if isinstance(value, xr.Dataset):
             xr_datasets.append(value)
+            
         # Raise warning for things that aren't xarrays/ints/floats/strings
         elif not isinstance(value, (int, float, str)):
             print(f'Warning: {key} not saved to .nc due to type {type(value)}')
